+++ src/mlpack/methods/range_search/range_search.hpp
@@ -29,47 +29,19 @@ class RangeSearch
 {
  public:
   /**
-   * Initialize the RangeSearch object with a different reference set and a
-   * query set.  Optionally, perform the computation in naive mode or
-   * single-tree mode, and set the leaf size used for tree-building.
-   * Additionally, an instantiated metric can be given, for cases where the
-   * distance metric holds data.
+   * Initialize the RangeSearch object with a given reference dataset (this is
+   * the dataset which is searched).  Optionally, perform the computation in
+   * naive mode or single-tree mode. Additionally, an instantiated metric can be
+   * given, for cases where the distance metric holds data.
    *
    * This method will copy the matrices to internal copies, which are rearranged
    * during tree-building.  You can avoid this extra copy by pre-constructing
    * the trees and passing them using a different constructor.
    *
    * @param referenceSet Reference dataset.
-   * @param querySet Query dataset.
    * @param naive Whether the computation should be done in O(n^2) naive mode.
    * @param singleMode Whether single-tree computation should be used (as
    *      opposed to dual-tree computation).
-   * @param leafSize The leaf size to be used during tree construction.
-   * @param metric Instantiated distance metric.
-   */
-  RangeSearch(const typename TreeType::Mat& referenceSet,
-              const typename TreeType::Mat& querySet,
-              const bool naive = false,
-              const bool singleMode = false,
-              const MetricType metric = MetricType());
-
-  /**
-   * Initialize the RangeSearch object with only a reference set, which will
-   * also be used as a query set.  Optionally, perform the computation in naive
-   * mode or single-tree mode, and set the leaf size used for tree-building.
-   * Additionally an instantiated metric can be given, for cases where the
-   * distance metric holds data.
-   *
-   * This method will copy the reference matrix to an internal copy, which is
-   * rearranged during tree-building.  You can avoid this extra copy by
-   * pre-constructing the reference tree and passing it using a different
-   * constructor.
-   *
-   * @param referenceSet Reference dataset.
-   * @param naive Whether the computation should be done in O(n^2) naive mode.
-   * @param singleMode Whether single-tree computation should be used (as
-   *      opposed to dual-tree computation).
-   * @param leafSize The leaf size to be used during tree construction.
    * @param metric Instantiated distance metric.
    */
   RangeSearch(const typename TreeType::Mat& referenceSet,
@@ -78,14 +50,12 @@ class RangeSearch
               const MetricType metric = MetricType());
 
   /**
-   * Initialize the RangeSearch object with the given datasets and
-   * pre-constructed trees.  It is assumed that the points in referenceSet and
-   * querySet correspond to the points in referenceTree and queryTree,
-   * respectively.  Optionally, choose to use single-tree mode.  Naive
-   * mode is not available as an option for this constructor; instead, to run
-   * naive computation, construct a tree with all the points in one leaf (i.e.
-   * leafSize = number of points).  Additionally, an instantiated distance
-   * metric can be given, for cases where the distance metric holds data.
+   * Initialize the RangeSearch object with the given pre-constructed reference
+   * tree (this is the tree built on the reference set, which is the set that is
+   * searched).  Optionally, choose to use single-tree mode, which will not
+   * build a tree on query points.  Naive mode is not available as an option for
+   * this constructor.  Additionally, an instantiated distance metric can be
+   * given, for cases where the distance metric holds data.
    *
    * There is no copying of the data matrices in this constructor (because
    * tree-building is not necessary), so this is the constructor to use when
@@ -93,45 +63,8 @@ class RangeSearch
    *
    * @note
    * Because tree-building (at least with BinarySpaceTree) modifies the ordering
-   * of a matrix, be sure you pass the modified matrix to this object!  In
-   * addition, mapping the points of the matrix back to their original indices
-   * is not done when this constructor is used.
-   * @endnote
-   *
-   * @param referenceTree Pre-built tree for reference points.
-   * @param queryTree Pre-built tree for query points.
-   * @param referenceSet Set of reference points corresponding to referenceTree.
-   * @param querySet Set of query points corresponding to queryTree.
-   * @param singleMode Whether single-tree computation should be used (as
-   *      opposed to dual-tree computation).
-   * @param metric Instantiated distance metric.
-   */
-  RangeSearch(TreeType* referenceTree,
-              TreeType* queryTree,
-              const typename TreeType::Mat& referenceSet,
-              const typename TreeType::Mat& querySet,
-              const bool singleMode = false,
-              const MetricType metric = MetricType());
-
-  /**
-   * Initialize the RangeSearch object with the given reference dataset and
-   * pre-constructed tree.  It is assumed that the points in referenceSet
-   * correspond to the points in referenceTree.  Optionally, choose to use
-   * single-tree mode.  Naive mode is not available as an option for this
-   * constructor; instead, to run naive computation, construct a tree with all
-   * the points in one leaf (i.e. leafSize = number of points).  Additionally,
-   * an instantiated distance metric can be given, for the case where the
-   * distance metric holds data.
-   *
-   * There is no copying of the data matrices in this constructor (because
-   * tree-building is not necessary), so this is the constructor to use when
-   * copies absolutely must be avoided.
-   *
-   * @note
-   * Because tree-building (at least with BinarySpaceTree) modifies the ordering
-   * of a matrix, be sure you pass the modified matrix to this object!  In
-   * addition, mapping the points of the matrix back to their original indices
-   * is not done when this constructor is used.
+   * of a matrix, be aware that mapping of the points back to their original
+   * indices is not done when this constructor is used.
    * @endnote
    *
    * @param referenceTree Pre-built tree for reference points.
@@ -141,7 +74,6 @@ class RangeSearch
    * @param metric Instantiated distance metric.
    */
   RangeSearch(TreeType* referenceTree,
-              const typename TreeType::Mat& referenceSet,
               const bool singleMode = false,
               const MetricType metric = MetricType());
 
@@ -152,11 +84,11 @@ class RangeSearch
   ~RangeSearch();
 
   /**
-   * Search for all points in the given range, returning the results in the
-   * neighbors and distances objects.  Each entry in the external vector
-   * corresponds to a query point.  Each of these entries holds a vector which
-   * contains the indices and distances of the reference points falling into the
-   * given range.
+   * Search for all reference points in the given range for each point in the
+   * query set, returning the results in the neighbors and distances objects.
+   * Each entry in the external vector corresponds to a query point.  Each of
+   * these entries holds a vector which contains the indices and distances of
+   * the reference points falling into the given range.
    *
    * That is:
    *
@@ -171,6 +103,83 @@ class RangeSearch
    *
    * - neighbors[i] and distances[i] are not sorted in any particular order.
    *
+   * @param querySet Set of query points to search with.
+   * @param range Range of distances in which to search.
+   * @param neighbors Object which will hold the list of neighbors for each
+   *      point which fell into the given range, for each query point.
+   * @param distances Object which will hold the list of distances for each
+   *      point which fell into the given range, for each query point.
+   */
+  void Search(const typename TreeType::Mat& querySet,
+              const math::Range& range,
+              std::vector<std::vector<size_t>>& neighbors,
+              std::vector<std::vector<double>>& distances);
+
+  /**
+   * Given a pre-built query tree, search for all reference points in the given
+   * range for each point in the query set, returning the results in the
+   * neighbors and distances objects.
+   *
+   * Each entry in the external vector corresponds to a query point.  Each of
+   * these entries holds a vector which contains the indices and distances of
+   * the reference points falling into the given range.
+   *
+   * That is:
+   *
+   * - neighbors.size() and distances.size() both equal the number of query
+   *   points.
+   *
+   * - neighbors[i] contains the indices of all the points in the reference set
+   *   which have distances inside the given range to query point i.
+   *
+   * - distances[i] contains all of the distances corresponding to the indices
+   *   contained in neighbors[i].
+   *
+   * - neighbors[i] and distances[i] are not sorted in any particular order.
+   *
+   * If either naive or singleMode are set to true, this will throw an
+   * invalid_argument exception; passing in a query tree implies dual-tree
+   * search.
+   *
+   * If you want to use the reference tree as the query tree, instead call the
+   * overload of Search() that does not take a query set.
+   *
+   * @param queryTree Tree built on query points.
+   * @param range Range of distances in which to search.
+   * @param neighbors Object which will hold the list of neighbors for each
+   *      point which fell into the given range, for each query point.
+   * @param distances Object which will hold the list of distances for each
+   *      point which fell into the given range, for each query point.
+   */
+  void Search(TreeType* queryTree,
+              const math::Range& range,
+              std::vector<std::vector<size_t>>& neighbors,
+              std::vector<std::vector<double>>& distances);
+
+  /**
+   * Search for all points in the given range for each point in the reference
+   * set (which was passed to the constructor), returning the results in the
+   * neighbors and distances objects.  This means that the query set and the
+   * reference set are the same.
+   *
+   * Each entry in the external vector corresponds to a query point.  Each of
+   * these entries holds a vector which contains the indices and distances of
+   * the reference points falling into the given range.
+   *
+   * That is:
+   *
+   * - neighbors.size() and distances.size() both equal the number of query
+   *   points.
+   *
+   * - neighbors[i] contains the indices of all the points in the reference set
+   *   which have distances inside the given range to query point i.
+   *
+   * - distances[i] contains all of the distances corresponding to the indices
+   *   contained in neighbors[i].
+   *
+   * - neighbors[i] and distances[i] are not sorted in any particular order.
+   *
+   * @param queryTree Tree built on query points.
    * @param range Range of distances in which to search.
    * @param neighbors Object which will hold the list of neighbors for each
    *      point which fell into the given range, for each query point.
@@ -178,8 +187,8 @@ class RangeSearch
    *      point which fell into the given range, for each query point.
    */
   void Search(const math::Range& range,
-              std::vector<std::vector<size_t> >& neighbors,
-              std::vector<std::vector<double> >& distances);
+              std::vector<std::vector<size_t>>& neighbors,
+              std::vector<std::vector<double>>& distances);
 
   // Returns a string representation of this object.
   std::string ToString() const;
@@ -187,29 +196,15 @@ class RangeSearch
  private:
   //! Copy of reference matrix; used when a tree is built internally.
   typename TreeType::Mat referenceCopy;
-  //! Copy of query matrix; used when a tree is built internally.
-  typename TreeType::Mat queryCopy;
-
   //! Reference set (data should be accessed using this).
   const typename TreeType::Mat& referenceSet;
-  //! Query set (data should be accessed using this).
-  const typename TreeType::Mat& querySet;
-
   //! Reference tree.
   TreeType* referenceTree;
-  //! Query tree (may be NULL).
-  TreeType* queryTree;
-
   //! Mappings to old reference indices (used when this object builds trees).
   std::vector<size_t> oldFromNewReferences;
-  //! Mappings to old query indices (used when this object builds trees).
-  std::vector<size_t> oldFromNewQueries;
 
   //! If true, this object is responsible for deleting the trees.
   bool treeOwner;
-  //! If true, a query set was passed; if false, the query set is the reference
-  //! set.
-  bool hasQuerySet;
 
   //! If true, O(n^2) naive computation is used.
   bool naive;
@@ -218,9 +213,6 @@ class RangeSearch
 
   //! Instantiated distance metric.
   MetricType metric;
-
-  //! The number of pruned nodes during computation.
-  size_t numPrunes;
 };
 
 }; // namespace range