+++ HISTORY.md
@@ -21,6 +21,9 @@
 
   * Add softmax regression, contributed by Siddharth Agrawal and QiaoAn Chen.
 
+  * Changed NeighborSearch and RangeSearch API; these classes now take the query
+    sets in the Search() method, instead of in the constructor.
+
 ### mlpack 1.0.11
 ###### 2014-12-11
   * Proper handling of dimension calculation in PCA.